Update TDA7439.h
Update definitions of variables to ensure they are 8bit - Compatibility with STM32 Arduino

+++ TDA7439.h
@@ -36,12 +36,12 @@ class TDA7439
 {
   public:
     TDA7439();
-	void setInput(int input); // 1 to 4
-	void inputGain(int gain); // 0 to 30
-	void setVolume(int volume); // 0 to 48 ( 0 is mute)
-	void setSnd(int val, int range); //-7 to +7 , 1 - Bass | 2 - Mids | 3 - Trebble
+	void setInput(uint8_t input); // 1 to 4
+	void inputGain(uint8_t gain); // 0 to 30
+	void setVolume(uint8_t volume); // 0 to 48 ( 0 is mute)
+	void setSnd(int8_t val, uint8_t range); //-7 to +7 , 1 - Bass | 2 - Mids | 3 - Trebble
 	void mute();
-	void spkAtt(int att); // Output attenuation 0 to 79 (db) can be used for balance but not enabled in this code.
+	void spkAtt(uint8_t att); // Output attenuation 0 to 79 (db) can be used for balance but not enabled in this code.
   private:
 	void writeWire(char a, char b);
 };
